pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "epok-toolkit"
version = "1.11.1"
description = "Una herramienta para la gestión de tareas y procesos en Django con Celery."
authors = [
    { name="Fernando Leon Franco", email="fernanlee2131@gmail.com" }
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Framework :: Django",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "Django>=5.0.2",
    "djangorestframework>=3.16.0",
    "celery[redis]>=5.5.3",
    "django-celery-beat>=2.6.0",
    "colorstreak>=0.1.0",
    "PyJWT>=2.10.1",
]

[tool.setuptools.packages.find]
where = ["."]